### 2024-04-16 #6630 @editorialbot
[PRE REVIEW]: Giga Connectome: a BIDS-app for time series and functional connectome extraction
**Submitting author:** @htwangtw (<a href="http://orcid.org/0000-0003-4078-2038">Hao-Ting Wang</a>)
**Repository:** https://github.com/bids-apps/giga_connectome
**Branch with paper.md** (empty if default branch):
**Version:** 0.5.0
