### 2024-04-16 #6630 @Kevin-Mattheus-Moerman
- [ ] Please study the above reference check :point_up: and see if you can address any of the reported potential DOI issues. If needed, you can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.
