### 2024-04-16 #6630 @htwangtw
Hi @Kevin-Mattheus-Moerman the flagged paper is a registered report hence there's no doi yet. I can ask the author to see what's possible.
