+++ paper/paper.bib
@@ -183,7 +183,8 @@
     title = {Investigating the convergence of resting-state functional connectivity profiles in {Alzheimer}'s disease with neuropsychiatric symptoms and schizophrenia},
     author = {Clarke, Natasha and Moreau, Clara A. and Harvey, Annabelle and Wang, Hao-Ting and Orban, Pierre and Zahinoor, Ismail and Jacquemont, Sébastien and Bellec, Pierre},
     year = {2024},
-    journal = {Brain Communications}
+    journal = {Brain Communications},
+    url = {http://osf.io/t7jkz}
 }
 
 
