+++ paper/paper.bib
@@ -184,7 +184,8 @@
     author = {Clarke, Natasha and Moreau, Clara A. and Harvey, Annabelle and Wang, Hao-Ting and Orban, Pierre and Zahinoor, Ismail and Jacquemont, Sébastien and Bellec, Pierre},
     year = {2024},
     journal = {Brain Communications},
-    url = {http://osf.io/t7jkz}
+    doi = {10.17605/OSF.IO/T7JKZ},
+    url = {http://osf.io/t7jkz},
 }
 
 
