### 2024-04-30 #6630 @Kevin-Mattheus-Moerman
@htwangtw yes we were waiting for the scope review. We've just concluded the review and I can share that is was determined this work may be in scope for JOSS. As such I will proceed to find a handing editor.
