### 2024-05-16 570be42
EHN save denoising details into meta data
