### 2024-05-17 94b998c
sampling frequency is a float
