### 2024-05-17 90549a9
Address case of sample mask is none
