### 2024-05-17 ced47b3
EHN save denoising details into meta data (#145)
* EHN save denoising details into meta data

* sampling frequency is a float

* FIX module import

* Address case of sample mask is none

* Add updates to doc

* fill up the docs
