+++ giga_connectome/__init__.py
@@ -6,17 +6,9 @@ try:
 except ImportError:
     pass
 
-from .atlas import load_atlas_setting
-from .denoise import get_denoise_strategy
-from .mask import generate_gm_mask_atlas
-from .postprocess import run_postprocessing_dataset
 
 __all__ = [
     "__copyright__",
     "__packagename__",
     "__version__",
-    "generate_gm_mask_atlas",
-    "load_atlas_setting",
-    "run_postprocessing_dataset",
-    "get_denoise_strategy",
 ]
