+++ giga_connectome/denoise.py
@@ -8,6 +8,7 @@ import numpy as np
 import pandas as pd
 from nibabel import Nifti1Image
 from nilearn.interfaces import fmriprep
+from nilearn.interfaces.fmriprep import load_confounds_utils as lc_utils
 from nilearn.maskers import NiftiMasker
 from pkg_resources import resource_filename
 
@@ -39,6 +40,17 @@ STRATEGY_TYPE = TypedDict(
         "name": str,
         "function": Callable[..., Any],
         "parameters": Dict[str, Union[str, List[str]]],
+    },
+)
+
+METADATA_TYPE = TypedDict(
+    "METADATA_TYPE",
+    {
+        "ConfoundRegressors": List[str],
+        "NumberOfVolumesDiscardedByMotionScrubbing": int,
+        "NumberOfVolumesDiscardedByNonsteadyStatesDetector": int,
+        "MeanFramewiseDisplacement": float,
+        "SamplingFrequency": float,
     },
 )
 
@@ -105,13 +117,65 @@ def is_ica_aroma(strategy: STRATEGY_TYPE) -> bool:
         raise ValueError(f"Invalid input dictionary. {strategy['parameters']}")
 
 
+def denoise_meta_data(strategy: STRATEGY_TYPE, img: str) -> METADATA_TYPE:
+    """Get metadata of the denoising process.
+
+    Including: column names of the confound regressors, number of
+    volumes discarded by motion scrubbing, number of volumes discarded
+    by non-steady states detector, mean framewise displacement and
+    place holder for sampling frequency (1/TR).
+
+    Parameters
+    ----------
+    strategy : dict
+        Denoising strategy parameter to pass to load_confounds_strategy
+        or load_confounds.
+    img : str
+        Path to the nifti image to denoise.
+
+    Returns
+    -------
+    dict
+        Metadata of the denoising process.
+    """
+    cf, sm = strategy["function"](img, **strategy["parameters"])
+    cf_file = lc_utils.get_confounds_file(img, flag_full_aroma=False)
+    framewise_displacement = pd.read_csv(cf_file, sep="\t")[
+        "framewise_displacement"
+    ]
+    mean_fd = np.mean(framewise_displacement)
+    # get non steady state volumes
+    _, sample_mask_non_steady = fmriprep.load_confounds(
+        img, strategy=["high_pass"]
+    )
+    n_non_steady = (
+        cf.shape[0] - sample_mask_non_steady.shape[0]
+        if sample_mask_non_steady is not None
+        else 0
+    )
+    n_scrub = 0
+    if "scrubbing" in strategy["parameters"].get(
+        "denoise_strategy", ""
+    ) or "srub" in strategy["parameters"].get("strategy", []):
+        n_scrub = cf.shape[0] - sm.shape[0] - n_non_steady
+
+    meta_data: METADATA_TYPE = {
+        "ConfoundRegressors": cf.columns.tolist(),
+        "NumberOfVolumesDiscardedByMotionScrubbing": n_scrub,
+        "NumberOfVolumesDiscardedByNonsteadyStatesDetector": n_non_steady,
+        "MeanFramewiseDisplacement": mean_fd,
+        "SamplingFrequency": np.nan,  # place holder
+    }
+    return meta_data
+
+
 def denoise_nifti_voxel(
     strategy: STRATEGY_TYPE,
     group_mask: str | Path,
     standardize: bool,
     smoothing_fwhm: float,
     img: str,
-) -> Nifti1Image:
+) -> Nifti1Image | None:
     """Denoise voxel level data per nifti image.
 
     Parameters
@@ -146,7 +210,6 @@ def denoise_nifti_voxel(
         standardize=standardize,
[…]
