+++ giga_connectome/tests/test_utils.py
@@ -33,10 +33,15 @@ def test_check_check_filter():
     assert "dseg" in str(msg.value)
 
 
-def test_parse_bids_name():
-    subject, session, specifier = utils.parse_bids_name(
-        "sub-01_ses-ah_task-rest_run-1_space-MNIfake_res-2_desc-preproc_bold.nii.gz"
-    )
+@pytest.mark.parametrize(
+    "source_file",
+    [
+        "sub-01_ses-ah_task-rest_run-1_space-MNIfake_res-2_desc-preproc_bold.nii.gz",
+        "sub-01_ses-ah_task-rest_run-1_space-MNIfake_res-2_desc-brain_mask.nii.gz",
+    ],
+)
+def test_parse_bids_name(source_file):
+    subject, session, specifier = utils.parse_bids_name(source_file)
     assert subject == "sub-01"
     assert session == "ses-ah"
     assert specifier == "ses-ah_task-rest_run-1"
@@ -53,3 +58,86 @@ def test_get_subject_lists():
     )
     assert len(subjects) == 1
     assert subjects[0] == "01"
+
+
+def test_output_filename():
+    source_file = "sub-01_ses-ah_task-rest_run-1_space-MNIfake_res-2_desc-preproc_bold.nii.gz"
+    atlas = "fake"
+    atlas_desc = "100"
+    strategy = "simple"
+
+    generated_target = utils.output_filename(
+        source_file=source_file,
+        atlas=atlas,
+        suffix="timeseries",
+        extension="tsv",
+        strategy=strategy,
+        atlas_desc=atlas_desc,
+    )
+    assert (
+        "sub-01_ses-ah_task-rest_run-1_seg-fake100_desc-denoiseSimple_timeseries.tsv"
+        == generated_target
+    )
+
+    generated_target = utils.output_filename(
+        source_file=source_file,
+        atlas=atlas,
+        suffix="timeseries",
+        extension="json",
+        strategy=strategy,
+        atlas_desc=atlas_desc,
+    )
+    assert (
+        "sub-01_ses-ah_task-rest_run-1_desc-denoiseSimple_timeseries.json"
+        == generated_target
+    )
+
+    generated_target = utils.output_filename(
+        source_file=source_file,
+        atlas=atlas,
+        suffix="relmat",
+        extension="tsv",
+        strategy=strategy,
+        atlas_desc=atlas_desc,
+    )
+    assert (
+        "sub-01_ses-ah_task-rest_run-1_seg-fake100_meas-PearsonCorrelation_desc-denoiseSimple_relmat.tsv"
+        == generated_target
+    )
+
+    generated_target = utils.output_filename(
+        source_file=source_file,
+        atlas=atlas,
+        suffix="report",
+        extension="html",
+        strategy=strategy,
+        atlas_desc=atlas_desc,
+    )
+    assert (
+        "sub-01_ses-ah_task-rest_run-1_seg-fake100_desc-denoiseSimple_report.html"
+        == generated_target
+    )
+
+    source_file = "sub-01_ses-ah_task-rest_run-1_space-MNIfake_res-2_desc-brain_mask.nii.gz"
+
+    generated_target = utils.output_filename(
+        source_file=source_file,
+        atlas=atlas,
+        suffix="dseg",
+        extension="nii.gz",
+        strategy="",
+        atlas_desc=atlas_desc,
+    )
+    assert "sub-01_seg-fake100_dseg.nii.gz" == generated_target
+
+    generated_target = utils.output_filename(
+        source_file=source_file,
+        atlas="",
+        suffix="mask",
+        extension="nii.gz",
+        strategy="",
+        atlas_desc="",
+    )
+    assert (
+        "sub-01_space-MNIfake_res-2_label-GM_mask.nii.gz" == generated_target
+    )
