+++ giga_connectome/atlas.py
@@ -94,60 +94,55 @@ def load_atlas_setting(
 
 
 def resample_atlas_collection(
-    template: str,
+    subject_seg_file_names: list[str],
     atlas_config: ATLAS_SETTING_TYPE,
-    group_mask_dir: Path,
-    group_mask: Nifti1Image,
+    subject_mask_dir: Path,
+    subject_mask: Nifti1Image,
 ) -> list[Path]:
     """Resample a atlas collection to group grey matter mask.
 
     Parameters
     ----------
-    template: str
-        Templateflow template name. This template should match the template of
-        `all_masks`.
+    subject_atlas_file_names: list of str
+        File names of subject atlas segmentations.
 
     atlas_config: dict
         Atlas name. Currently support Schaefer20187Networks, MIST, DiFuMo.
 
-    group_mask_dir: pathlib.Path
+    subject_mask_dir: pathlib.Path
         Path to where the outputs are saved.
 
-    group_mask : nibabel.nifti1.Nifti1Image
-        EPI (grey matter) mask for the current group of subjects.
+    subject_mask : nibabel.nifti1.Nifti1Image
+        EPI (grey matter) mask for the subject.
 
     Returns
     -------
     list of pathlib.Path
-        Paths to atlases sampled to group level grey matter mask.
+        Paths to subject specific segmentations created from atlases sampled
+        to individual grey matter mask.
     """
     gc_log.info("Resample atlas to group grey matter mask.")
-    resampled_atlases = []
+    subject_seg = []
 
     with progress_bar(text="Resampling atlases") as progress:
         task = progress.add_task(
             description="resampling", total=len(atlas_config["file_paths"])
         )
 
-        for desc in atlas_config["file_paths"]:
+        for seg_file, desc in zip(
+            subject_seg_file_names, atlas_config["file_paths"]
+        ):
             parcellation = atlas_config["file_paths"][desc]
             parcellation_resampled = resample_to_img(
-                parcellation, group_mask, interpolation="nearest"
+                parcellation, subject_mask, interpolation="nearest"
             )
-            filename = (
-                f"tpl-{template}_"
-                f"atlas-{atlas_config['name']}_"
-                "res-dataset_"
-                f"desc-{desc}_"
-                f"{atlas_config['type']}.nii.gz"
-            )
-            save_path = group_mask_dir / filename
+            save_path = subject_mask_dir / seg_file
             nib.save(parcellation_resampled, save_path)
-            resampled_atlases.append(save_path)
+            subject_seg.append(save_path)
 
         progress.update(task, advance=1)
 
-    return resampled_atlases
+    return subject_seg
 
 
 def get_atlas_labels() -> List[str]:

[…]
