### 2024-05-17 04b0be2
parametise the output name test
