+++ giga_connectome/tests/test_atlas.py
@@ -2,7 +2,7 @@ from giga_connectome.atlas import load_atlas_setting
 
 
 def test_load_atlas_setting():
-    atlas_config = load_atlas_setting("Schaefer20187Networks")
-    assert atlas_config["name"] == "Schaefer20187Networks"
+    atlas_config = load_atlas_setting("Schaefer2018")
+    assert atlas_config["name"] == "Schaefer2018"
     atlas_config = load_atlas_setting("HarvardOxfordCortical")
     assert atlas_config["name"] == "HarvardOxfordCortical"
