### 2024-05-17 c77bc00
Update documentations related to introducing seg entity
