+++ docs/source/outputs.md
@@ -16,12 +16,12 @@ the output will be save in `sub-<participant_id>/[ses-<ses_id>]/func`.
 For each input image (that is, preprocessed BOLD time series)
 and each atlas the following data files will be generated
 
-- a `[matches]_atlas-{atlas}_meas-PearsonCorrelation_desc-{atlas_description}{denoise_strategy}_relmat.tsv`
+- a `[matches]_seg-{atlas}{atlas_description}_meas-PearsonCorrelation_desc-denoise{denoise_strategy}_relmat.tsv`
   file that contains the correlation matrix between all the regions of the atlas
-- a `[matches]_atlas-{atlas}_desc-{atlas description}{denoise_strategy}_timeseries.tsv`
+- a `[matches]_seg-{atlas}{atlas_description}_desc-denoise{denoise_strategy}_timeseries.tsv`
   file that contains the extracted timeseries for each region of the atlas
 
-- `{atlas}` refers to the name of the atlas used (for example, `Schaefer20187Networks`)
+- `{atlas}` refers to the name of the atlas used (for example, `Schaefer2018`)
 - `{atlas_description}` refers to the sub type of atlas used (for example, `100Parcels7Networks`)
 - `{denoise_strategy}` refers to the denoise strategy passed to the command line
 
@@ -31,7 +31,7 @@ A JSON file is generated in the root of the output dataset (`meas-PearsonCorrela
 that contains metadata applicable to all `relmat.tsv` files.
 
 For each input image (that is, preprocessed BOLD time series)
-a `[matches]_atlas-{atlas}_timeseries.json`
+a `[matches]_desc-denoise{denoise_strategy}_timeseries.json`
 
 ### Example
 
@@ -43,51 +43,100 @@ a `[matches]_atlas-{atlas}_timeseries.json`
 ├── sub-1
 │   ├── ses-timepoint1
 │   │   └── func
-│   │       ├── sub-1_ses-timepoint1_task-probabilisticclassification_run-01_space-MNI152NLin2009cAsym_res-2_atlas-Schaefer20187Networks_desc-100Parcels7NetworksSimple_timeseries.tsv
-│   │       ├── sub-1_ses-timepoint1_task-probabilisticclassification_run-01_space-MNI152NLin2009cAsym_res-2_atlas-Schaefer20187Networks_desc-200Parcels7NetworksSimple_timeseries.tsv
-│   │       ├── sub-1_ses-timepoint1_task-probabilisticclassification_run-01_space-MNI152NLin2009cAsym_res-2_atlas-Schaefer20187Networks_meas-PearsonCorrelation_desc-100Parcels7NetworksSimple_relmat.tsv
-│   │       ├── sub-1_ses-timepoint1_task-probabilisticclassification_run-01_space-MNI152NLin2009cAsym_res-2_atlas-Schaefer20187Networks_meas-PearsonCorrelation_desc-200Parcels7NetworksSimple_relmat.tsv
-│   │       ├── sub-1_ses-timepoint1_task-probabilisticclassification_run-01_space-MNI152NLin2009cAsym_res-2_atlas-Schaefer20187Networks_timeseries.json
-│   │       ├── sub-1_ses-timepoint1_task-probabilisticclassification_run-02_space-MNI152NLin2009cAsym_res-2_atlas-Schaefer20187Networks_desc-100Parcels7NetworksSimple_timeseries.tsv
[…]
