### 2024-05-17 acc0c11
Reflect the change of Schaefer atlas name in github workflow
