### 2024-05-27 777f0e0
[BOT] update pre-commit hooks (#148)
* pre-commit hooks auto-update

* FIX spelling suggested by precommit

---------

Co-authored-by: htwangtw <13743617+htwangtw@users.noreply.github.com>
Co-authored-by: Hao-Ting Wang <htwangtw@gmail.com>
