### 2024-05-28 70f9ddc
FIX type of the new deprication action class
