### 2024-05-28 04f0669
[full_test] deprecation warning for atlas name
