### 2024-05-28 63aa083
[full_test] check type of potentially depricated atlas input
