### 2024-05-28 a6f218b
[full_test] TEST capture log correctly
