### 2024-05-28 3b38814
add deprication related details to doc
