### 2024-05-28 692948a
Update subject segmentation outputs (#146)
* Update subject segmentation outputs

* desc -> atlas_desc to make the source explicit

* parametise the output name test

* ensure the file name changes are reflected in the workflow

* FIX test should reflect new changw

* Update documentations related to introducing seg entity

* Reflect the change of Schaefer atlas name in github workflow

* [full_test] test deprication

* FIX type of the new deprication action class

* [full_test] deprecation warning for atlas name

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, se
[…]
