+++ giga_connectome/tests/test_utils.py
@@ -33,10 +33,15 @@ def test_check_check_filter():
     assert "dseg" in str(msg.value)
 
 
-def test_parse_bids_name():
-    subject, session, specifier = utils.parse_bids_name(
-        "sub-01_ses-ah_task-rest_run-1_space-MNIfake_res-2_desc-preproc_bold.nii.gz"
-    )
+@pytest.mark.parametrize(
+    "source_file",
+    [
+        "sub-01_ses-ah_task-rest_run-1_space-MNIfake_res-2_desc-preproc_bold.nii.gz",
+        "sub-01_ses-ah_task-rest_run-1_space-MNIfake_res-2_desc-brain_mask.nii.gz",
+    ],
+)
+def test_parse_bids_name(source_file):
+    subject, session, specifier = utils.parse_bids_name(source_file)
     assert subject == "sub-01"
     assert session == "ses-ah"
     assert specifier == "ses-ah_task-rest_run-1"
@@ -53,3 +58,62 @@ def test_get_subject_lists():
     )
     assert len(subjects) == 1
     assert subjects[0] == "01"
+
+
+@pytest.mark.parametrize(
+    "suffix,extension,target",
+    [
+        (
+            "timeseries",
+            "tsv",
+            "sub-01_ses-ah_task-rest_run-1_seg-fake100_desc-denoiseSimple_timeseries.tsv",
+        ),
+        (
+            "timeseries",
+            "json",
+            "sub-01_ses-ah_task-rest_run-1_desc-denoiseSimple_timeseries.json",
+        ),
+        (
+            "relmat",
+            "tsv",
+            "sub-01_ses-ah_task-rest_run-1_seg-fake100_meas-PearsonCorrelation_desc-denoiseSimple_relmat.tsv",
+        ),
+        (
+            "report",
+            "html",
+            "sub-01_ses-ah_task-rest_run-1_seg-fake100_desc-denoiseSimple_report.html",
+        ),
+    ],
+)
+def test_output_filename(suffix, extension, target):
+    source_file = "sub-01_ses-ah_task-rest_run-1_space-MNIfake_res-2_desc-preproc_bold.nii.gz"
+
+    generated_target = utils.output_filename(
+        source_file=source_file,
+        atlas="fake",
+        suffix=suffix,
+        extension=extension,
+        strategy="simple",
+        atlas_desc="100",
+    )
+    assert target == generated_target
+
+
+@pytest.mark.parametrize(
+    "atlas,atlas_desc,suffix,target",
+    [
+        ("fake", "100", "dseg", "sub-01_seg-fake100_dseg.nii.gz"),
+        ("", "", "mask", "sub-01_space-MNIfake_res-2_label-GM_mask.nii.gz"),
+    ],
+)
+def test_output_filename_seg(atlas, atlas_desc, suffix, target):
+    source_file = "sub-01_ses-ah_task-rest_run-1_space-MNIfake_res-2_desc-brain_mask.nii.gz"
+    generated_target = utils.output_filename(
+        source_file=source_file,
+        atlas=atlas,
+        suffix=suffix,
+        extension="nii.gz",
+        strategy="",
+        atlas_desc=atlas_desc,
+    )
+    assert target == generated_target

[…]
