### 2024-05-30 #6630 @htwangtw
People on [BEP 17](https://docs.google.com/document/d/1ugBdUF6dhElXdj3u9vw0iWjE6f_Bibsro3ah7sRV0GA/edit#heading=h.mqkmyp254xh6) or [BEP12](https://github.com/bids-standard/bids-specification/pull/519), [BEP38](https://docs.google.com/document/d/1RxW4cARr3-EiBEcXjLpSIVidvnUSHE7yJCUY91i5TfM/edit)
would all be good candidates and I genuinely want their feedback on the output data structures.

I would like peerherholz to review this but I will leave it up to him to see if this is suitable.
