### 2024-05-31 #6630 @PeerHerholz
Hi folks,

thanks for the offer.

I would love to review this submission but unfortunately, my schedule is rather packed until mid-July and thus,
I won't have to do the work this deserves. However, @htwangtw if you want to chat about output structures and metadata, please let me know. I should definitely have 1-2 h for this.

Thanks and sorry again.

Best, Peer
