### 2024-05-31 4772ec8
FIX don't redefine subject_mask_nii
