### 2024-05-31 fb3069b
cover the mask reuse case in smoke test
