### 2024-05-31 abb2407
FIX reuse pregenerated subject segmentation or masks (#150)
* FIX reuse pregenerated subject segmentation or masks

* FIX don't redefine subject_mask_nii

* cover the mask reuse case in smoke test
