### 2024-05-31 8e8be89
Fix/reuse subj mask (#152)
* imporve test converate

* full_test
