### 2024-05-31 6af3eea
[full_test] fix rerun
