### 2024-05-31 f413d03
[full_test] fix output of mask/seg checking
