### 2024-05-31 e44fca6
When re-runing with the same atlases, load segmentations (#153)
* [full_test] fix rerun

* [full_test] fix output of mask/seg checking
