### 2024-06-03 6a71292
FIX Address the case of no volumes to scrub when given the scrubbing strategy
