### 2024-06-03 f708610
FIX Address the case of no volumes to scrub when given the scrubbing strategy (#156)
