+++ .pre-commit-config.yaml
@@ -11,7 +11,7 @@ repos:
     -   id: check-added-large-files
         args: ['--maxkb=1000']
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.10.0
+    rev: v1.10.1
     hooks:
     -   id: mypy
         additional_dependencies: [types-all, pandas-stubs, types-tqdm]
@@ -29,7 +29,7 @@ repos:
         args: ["--skip=*.svg"]
         exclude: 'paper/.*'
 -   repo: https://github.com/PyCQA/flake8
-    rev: 7.0.0
+    rev: 7.1.0
     hooks:
     -   id: flake8
 -   repo: https://github.com/hadolint/hadolint
