+++ .pre-commit-config.yaml
@@ -11,7 +11,7 @@ repos:
     -   id: check-added-large-files
         args: ['--maxkb=1000']
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.10.1
+    rev: v1.11.0
     hooks:
     -   id: mypy
         additional_dependencies: [types-all, pandas-stubs, types-tqdm]
