### 2024-07-30 #7061 @editorialbot
[REVIEW]: Giga Connectome: a BIDS-app for time series and functional connectome extraction
**Submitting author:** @htwangtw (<a href="http://orcid.org/0000-0003-4078-2038">Hao-Ting Wang</a>)
**Repository:** https://github.com/bids-apps/giga_connectome
**Branch with paper.md** (empty if default branch):
**Version:** 0.6.0
