### 2024-07-30 #7061 @mstimberg
at the top of a new comment in this thread.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/7061#issue-2438189455).
