### 2024-07-31 bids-apps/giga_connectome#161 @shnizzedy
Document contributors and their roles
### Your idea

As part of https://github.com/openjournals/joss-reviews/issues/7061#issuecomment-2258789463, I'm asked to verify
> Does the full list of paper authors seem appropriate and complete?

The answer to that question isn't easily apparent to me. For the _JOSS_ submission and generally, having the contributors and their roles clearly documented would be helpful.

I can see from [this repo's contributor graph](https://github.com/bids-apps/giga_connectome/graphs/contributors) that Hao-Ting and Rémi contributed significantly to the code, but it's unclear what other roles they and the other authors contributed (or if that information is documented, it's not particularly easy to find).

For the purposes of the review, I think just responding to this issue is probably sufficient, but using something like [All Contributors](https://allcontributors.org/docs/en/overview) or [CRediT](https://credit.niso.org/contributor-roles-defined/) in the README and/or documentation would be a more sustainable (and broadly useful) solution.
