### 2024-08-01 #7061 @shnizzedy
@mstimberg I was unaware of this work prior to being invited to review. I think I am able to make an impartial assessment of this work and request that these conflicts be waived:

### Conflicts of interest

- [ ] Rémi Gau and I are co-authors (4th and 11th, respectively, of 19 co-authors) on a currently-in-review paper[^1] describing a metadata project[^2] that we both contributed to.
- [ ] Rémi Gau, Pierre Bellec and I were all involved in getting the NMIND Consortium[^3]<sup>,</sup>[^4] (devising best practices to improve reproducibility and reduce duplicated effort) off the ground.
- [ ] Rémi Gau and Pierre Bellec have both published papers[^5]<sup>,</sup>[^6] describing Brainhack, a recurring workshop with a New York City event hosted annually by my employer, the Child Mind Institute. All of the authors and myself are members of the Brainhack community with profiles on the Brainhack Mattermost[^7].
- [ ] I'm a co-author of a pre-print[^8] cited in this manuscript[^9].

### References
[^1]: Chen Y, Jarecka D, Abraham SA, **Gau R**, Ng E, Low DM, Bevers I, Johnson A, Keshavan A, Klein A, **Clucas J**, Rosli Z, Hodge SM, Linkersdörfer J, Bartsch H, Das S, Fair D, Kennedy D, Ghosh SS. (2024). ReproSchema: Enhancing Research Reproducibility through Standardized Survey Data Collection. _JMIR Preprints. 21/06/2024_:63343. DOI:[10.2196/preprints.63343](https://doi.org/10.2196/preprints.63343).
[^2]: Reproschema Contributors. (2020). ReproNim/reproschema: 1.0.0 Release Candidate 2 (1.0.0-rc2). _Zenodo_. DOI:[10.5281/zenodo.4064940](https://doi.org/10.5281/zenodo.4064940).
[^3]: [NMIND](https://www.nmind.org).
[^4]: Kiar G, **Clucas J**, Feczko E, et al. (2023). Align with the NMIND consortium for better neuroimaging. _Nat Hum Behav 7_, 1027–1028. DOI:[10.1038/s41562-023-01647-0](https://doi.org/10.1038/s41562-023-01647-0).
[^5]: **Gau R**, Noble S, Heuer K, et al. (2021). Brainhack: Developing a culture of open, inclusive, community-driven neuroscience. _Neuroview 109_, 11, 1769-1775. DOI:[10.1016/j.neuron.2021.04.001](https://doi.org/10.1016/j.neuron.2021.04.001).
[^6]: Craddock RC, Margulies DS, **Bellec P**, et al. (2016). Brainhack: a collaborative workshop for the open neuroscience community. _GigaScience 5_, 1. DOI:[10.1186/s13742-016-0121-x](https://doi.org/10.1186/s13742-016-0121-x).
[^7]: [Brainhack Mattermost](https://mattermost.brainhack.org/brainhack).<br/>![Hao-Ting Wang](https://github.com/user-attachments/assets/0e3e9ad1-8824-4553-b3a3-089644a8b887) ![Rémi Gau](https://github.com/user-attachments/assets/c51e1c1a-f133-40ae-b7e6-4447c9d2c9de) ![Natasha Clarke](https://github.com/user-attachments/assets/88f464b0-5cfd-4720-82d9-7413ad734d12) ![Quentin Dessain](https://github.com/user-attachments/assets/646f865b-08b2-4cbd-825d-8638e8471d65)
![Pierre Bellec](https://github.com/user-attachments/assets/2de88537-a911-473c-a644-66879a365558)
![Jon Clucas](https://github.com/user-attachments/assets/a8e77d47-9856-4ed4-a242-61b26075f935)
[^8]: Li X, Esper NB, Ai L, Giavasis S, Jin H, Feczko E, Xu T, **Clucas J**, et al. (2024). Moving Beyond Processing and Analysis-Related Variation in Neuroscience. _bioRxiv 2021.12.01.470790_. DOI:[10.1101/2021.12.01.470790](https://doi.org/10.1101/2021.12.01.470790).
[^9]: [BibTeX entry](https://github.com/bids-apps/giga_connectome/blob/f67c572238cbb1354e8ff776f013a13f4a51f8a7/paper/paper.bib#L208-L217) cited on line 62 in [Giga Connectome: a BIDS-app for time series and functional connectome extraction](https://github.com/openjournals/joss-papers/blob/joss.07061/joss.07061/10.21105.joss.07061.pdf).
