### 2024-08-01 #7061 @mstimberg
@shnizzedy Many thanks for the detailed information. From my side, I'd say that the last point is not an issue (it regularly happens that reviewers are authors of software in the same field that get cited by the reviewed manuscript). I am also not worried about all of you being part of the Brainhack community – reviewers and authors being part of the same community is difficult to avoid if we want to have competent reviewers :blush:
The first two points are less clear, but again, I'd waive that COI. From what I see, a lot of people in the field are in some way part of the BIDS community, so these kinds of interactions seem to be hard to avoid. However, I'd be more comfortable if the track editor signed off on this as well. @Kevin-Mattheus-Moerman, what do you think?

CC @sappelhoff who is both a JOSS editor and a BIDS maintainer, so he might have valuable insights to add.
