### 2024-08-01 #7061 @sappelhoff
> From what I see, a lot of people in the field are in some way part of the BIDS community, so these kinds of interactions seem to be hard to avoid.

I agree ✅

I believe the COI for @shnizzedy can be waived.
