### 2024-08-02 bids-apps/giga_connectome#161 @htwangtw
Thanks for suggesting @shnizzedy! All contributors would be a good addition indeed.

QD contributed on the code to containerised the BIDS app.
NC tested the software by deploying it on 10+ datasets, caught bugs, and helped improved the features (see issue #36 #54 #106 for examples)
PB provided direction of design, such as adopting BIDS convention, and scope of the features.
