+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
     rev: v1.11.1
     hooks:
     -   id: mypy
-        additional_dependencies: [types-all, pandas-stubs, types-tqdm]
+        additional_dependencies: [pandas-stubs, types-tqdm, types-setuptools, types-Jinja2]
         args: [--config-file=pyproject.toml]
 -   repo: https://github.com/psf/black-pre-commit-mirror
     rev: 24.4.2
