### 2024-08-09 bids-apps/giga_connectome#166 @shnizzedy
Add examples
### Related to

Related to https://github.com/openjournals/joss-reviews/issues/7061

### Rationale

#### Clarify input data requirements

I'm unclear on what is necessary / required for the input data for `giga_connectome`. I'm able to successfully run the main branch and `unstable` Docker container with the example data from https://github.com/bids-apps/giga_connectome/blob/fb6f5d1ae8eec6d9a91b7aeaf57c96968adeaaae/tox.ini#L26-L45, but with [this filetree of fMRIPrep outputs](https://github.com/bids-apps/giga_connectome/issues/164#filetree) from another project, I'm getting
```Python
Traceback (most recent call last):
  File "/Users/jon.clucas/opt/mambaforge/envs/giga_connectome3.12/bin/giga_connectome", line 8, in <module>
    sys.exit(main())
             ^^^^^^
  File "/Users/jon.clucas/opt/mambaforge/envs/giga_connectome3.12/lib/python3.12/site-packages/giga_connectome/run.py", line 160, in main
    workflow(args)
  File "/Users/jon.clucas/opt/mambaforge/envs/giga_connectome3.12/lib/python3.12/site-packages/giga_connectome/workflow.py", line 82, in workflow
    subject_mask_nii, subject_seg_niis = generate_gm_mask_atlas(
                                         ^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/jon.clucas/opt/mambaforge/envs/giga_connectome3.12/lib/python3.12/site-packages/giga_connectome/mask.py", line 38, in generate_gm_mask_atlas
    subject, _, _ = utils.parse_bids_name(masks[0].path)
                                          ~~~~~^^^
IndexError: list index out of range
```

and it's not obvious to me what the discrepancy is.

#### JOSS review criteria

> ##### [Example usage](https://joss.readthedocs.io/en/latest/review_criteria.html#example-usage)
> The authors should include examples of how to use the software (ideally to solve real-world analysis problems).

― https://joss.readthedocs.io/en/latest/review_criteria.html#example-usage

The [current "Usage Notes"](https://giga-connectome.readthedocs.io/en/0.5.0/usage.html) in the `giga_connectome` documentation are mostly hypothetical, like

```bash
FMRIPREP_DIR=/path/to/fmriprep_output
OUTPUT_DIR=/path/to/connectom_output
WORKING_DIR=/path/to/working_dir
DENOISE_CONFIG=/path/to/denoise_config.json

GIGA_CONNECTOME=/path/to/giga-connectome.simg

apptainer run \
    --bind ${FMRIPREP_DIR}:/data/input \
    --bind ${OUTPUT_DIR}:/data/output \
    --bind ${WORKING_DIR}:/data/working \
    --bind ${DENOISE_CONFIG}:/data/denoise_config.json \
    ${GIGA_CONNECTOME} \
    -w /data/working \
    --denoise-strategy /data/denoise_config.json \
    /data/input \
    /data/output \
    participant
```

Even if it's just with the data used for testing, a real example would be helpful.
