+++ .pre-commit-config.yaml
@@ -17,7 +17,7 @@ repos:
         additional_dependencies: [pandas-stubs, types-tqdm, types-setuptools, types-Jinja2]
         args: [--config-file=pyproject.toml]
 -   repo: https://github.com/psf/black-pre-commit-mirror
-    rev: 24.4.2
+    rev: 24.8.0
     hooks:
     -   id: black
 -   repo: https://github.com/codespell-project/codespell
