### 2024-08-13 df40338
print version in CI
