### 2024-09-01 eb9a18f
[FIX] Make sure version docker images matches version of package in the image (#169)
* update docker ignore

* update docker ingore

* speed up CLI

* print version in CI
