### 2024-09-03 bids-apps/giga_connectome#171 @emullier
Unclear inputs required
In the context of the review [(https://github.com/openjournals/joss-reviews/issues/7061 )], I am trying to run the giga connectome on one publicly available subject ([(https://zenodo.org/records/5790821)])

It is not clear to me what inputs I should use to run the giga_connectome

I followed the following steps:

1. I downloaded the subject, passed it through the bids validator (1.14.0) without any error

2. I ran fmriprep (via docker version  24.1.0rc0), with all my fmriprep outputs generated in my derivatives directory
![image](https://github.com/user-attachments/assets/244f16ef-c286-427b-87e2-8ba1a4b04c43)

3. Finally I tried to run giga_connectome via this command line
```
docker run -it --rm -v  /home/localadmin/DATA/ds-sample_GigaConnectome/derivatives:/bids_dir -v  /home/localadmin/DATA/ds-sample_GigaConnectome/derivatives/giga_connectome:/out bids/giga_connectome bids_dir out  participant --participant_label 01
```

This is not clear for me if bids_dir should be my derivatives with the fmriprep output or my original bids directory?
I tried both but in both case I get the same error which seems to be a bids issue

````
          INFO     Indexing BIDS directory:                      workflow.py:66
                            bids_dir
Traceback (most recent call last):
  File "/usr/local/bin/giga_connectome", line 8, in <module>
    sys.exit(main())
  File "/usr/local/lib/python3.9/site-packages/giga_connectome/run.py", line 127, in main
    workflow(args)
  File "/usr/local/lib/python3.9/site-packages/giga_connectome/workflow.py", line 81, in workflow
    subj_data, _ = utils.get_bids_images(
  File "/usr/local/lib/python3.9/site-packages/giga_connectome/utils.py", line 81, in get_bids_images
    subj_data = {
  File "/usr/local/lib/python3.9/site-packages/giga_connectome/utils.py", line 82, in <dictcomp>
    dtype: layout.get(**layout_get_kwargs, **query)
  File "/usr/local/lib/python3.9/site-packages/bids/layout/layout.py", line 629, in get
    raise ValueError(msg + "If you're sure you want to impose "
ValueError: 'desc' is not a recognized entity. If you're sure you want to impose this constraint, set invalid_filters='allow'.

`````
