### 2024-09-03 #7061 @emullier
Hi @mstimberg, I started to play with the software 2 days ago, I currently have some issues running it on my subject
I opened two issues about my problems in running it on my subject (issue #171) and (issue #172).
Maybe if I could have some feedback on this?
