### 2024-09-03 #7061 @htwangtw
> Hi @jdkent and @emullier, did you already have a chance to have a closer look at the software and paper?
>
> Please be reminded of our review recommendations:
>
> > Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the [target repository](https://github.com/bids-apps/giga_connectome) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)
>
> @htwangtw: please have another look at the issues opened by @shnizzedy and update their status (e.g. close them) if necessary 🙏

Thanks for the reminder - I will get back on to this soon!
