### 2024-09-03 bids-apps/giga_connectome#171 @htwangtw
@emullier
Thanks for opening the issue!

I see that you are using the latest fMRIPrep release candidate version, which is not tested on.
All the developments of the software is aiming to support fMRIPrep LTS versions, sorry that this is not clear in our documentations. If you can preprocess the subject with the [latest LTS version](https://fmriprep.org/en/stable/changes.html#id1) and report back, that would be extremely helpful.

To help me debug this specific message, can you print out the tree of `sub-01/`? This will help me to understand if this is a issue stemmed from a new version or a bug...
