### 2024-09-03 bids-apps/giga_connectome#166 @shnizzedy
Related: https://github.com/bids-apps/giga_connectome/issues/171
