### 2024-09-05 #7061 @Kevin-Mattheus-Moerman
@mstimberg I think it would be best to remove @shnizzedy as a reviewer. There does seem to be either a potential or perceived COI, especially relating to the shared authorship and consortium membership which show a prolonged collaboration.

@shnizzedy thanks for sharing this information, and thanks for offering to help. Unfortunately our policies are quite strict on potential/perceived COIs. Hence we will proceed to remove you as reviewer. We do hope we may ask for your assistance for a different JOSS submission in the future. Thanks agian.
