### 2024-09-05 #7061 @mstimberg
Dear @shnizzedy, apologies that this took so long, and even more apologies that the final decision is contrary to what I indicated earlier. As @Kevin-Mattheus-Moerman explained above, we finally consider your closeness to some of the authors as being a potential COI – or at least perceivable as one. I therefore regret that I'll have to officially remove you as a reviewer. Thanks a million for your involvement, though, and I'd be very happy to see you come back to JOSS as a reviewer (or author) another time.

@htwangtw Of course, this does not mean that you have to discard the issues opened by @shnizzedy, please feel free to address them as you see fit.
