### 2024-09-20 bids-apps/giga_connectome#171 @Remi-Gau
- I think that this PR may fix that: https://github.com/bids-apps/giga_connectome/pull/174

I was getting the same "desc" not recognized issue
