### 2024-09-20 4c2d14e
set pybids config (#174)
