### 2024-09-26 #7061 @mstimberg
:wave: @emullier did you find time to have another look at the software? Please note that the authors have replied in the two issues you opened (giga_connectome/issues#171 and giga_connectome/issues#171).

@jdkent Can you please give us an estimate when you will be able to review this submission?
