### 2024-10-07 5b53dc8
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pre-commit/pre-commit-hooks: v4.6.0 → v5.0.0](https://github.com/pre-commit/pre-commit-hooks/compare/v4.6.0...v5.0.0)
- [github.com/psf/black-pre-commit-mirror: 24.8.0 → 24.10.0](https://github.com/psf/black-pre-commit-mirror/compare/24.8.0...24.10.0)
- [github.com/hadolint/hadolint: v2.13.0-beta → v2.13.1-beta](https://github.com/hadolint/hadolint/compare/v2.13.0-beta...v2.13.1-beta)
