+++ .pre-commit-config.yaml
@@ -2,7 +2,7 @@
 # See https://pre-commit.com/hooks.html for more hooks
 repos:
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.6.0
+    rev: v5.0.0
     hooks:
     -   id: trailing-whitespace
     -   id: end-of-file-fixer
@@ -17,7 +17,7 @@ repos:
         additional_dependencies: [pandas-stubs, types-tqdm, types-setuptools, types-Jinja2]
         args: [--config-file=pyproject.toml]
 -   repo: https://github.com/psf/black-pre-commit-mirror
-    rev: 24.8.0
+    rev: 24.10.0
     hooks:
     -   id: black
 -   repo: https://github.com/codespell-project/codespell
@@ -33,7 +33,7 @@ repos:
     hooks:
     -   id: flake8
 -   repo: https://github.com/hadolint/hadolint
-    rev: v2.13.0-beta
+    rev: v2.13.1-beta
     hooks:
     -   id: hadolint-docker
         name: Lint Dockerfiles
