### 2024-10-08 18e68d4
[MAINT] drop python 3.8 and support 3.13 (#179)
* drop 3.8 and support 3.13

* fix
