+++ pyproject.toml
@@ -6,18 +6,18 @@ build-backend = "hatchling.build"
 name = "giga_connectome"
 description = "Generate connectome from fMRIPrep outputs"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = { file="LICENSE" }
 authors = [
   { name="Hao-Ting Wang", email="htwangtw@gmail.com" },
 ]
 classifiers = [
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
 ]
 dependencies = [
   "nilearn[plotting] >=0.10.3",
