### 2024-10-15 #7061 @mstimberg
@emullier & @jdkent, could you please give us an update ?

> 👋 @emullier did you find time to have another look at the software? Please note that the authors have replied in the two issues you opened ([bids-apps/giga_connectome#171](https://github.com/bids-apps/giga_connectome/issues/171) and [bids-apps/giga_connectome#171](https://github.com/bids-apps/giga_connectome/issues/171)).
>
> @jdkent Can you please give us an estimate when you will be able to review this submission?
