+++ .pre-commit-config.yaml
@@ -11,7 +11,7 @@ repos:
     -   id: check-added-large-files
         args: ['--maxkb=1000']
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.11.2
+    rev: v1.12.1
     hooks:
     -   id: mypy
         additional_dependencies: [pandas-stubs, types-tqdm, types-setuptools, types-Jinja2]
