### 2024-12-01 bids-apps/giga_connectome#184 @jdkent
[BUG] dev installation types-all resolver unsatisfiable
### What version of the bids app were you using?

main branch

### Describe your problem in detail.

I'm using `uv` to install this project, and I cannot install a dev version of the project. I'm using python 3.10 specifically.

### What command did you run?

`uv pip install -e .[dev]`

I got this output:
```
× No solution found when resolving dependencies:
  ╰─▶ Because only the following versions of types-pkg-resources are available:
          types-pkg-resources==0.1.0
          types-pkg-resources==0.1.1
          types-pkg-resources==0.1.2
          types-pkg-resources==0.1.3
      and types-pkg-resources<=0.1.2 was yanked (reason: Use types-setuptools instead), we
      can conclude that types-pkg-resources<0.1.1 cannot be used.
      And because types-pkg-resources==0.1.3 was yanked (reason: Use types-setuptools
      instead) and types-all==1.0.0 depends on types-pkg-resources, we can conclude that
      types-all==1.0.0 cannot be used.
      And because only types-all==1.0.0 is available and
      giga-connectome[dev]==0.1.dev198+gf6e41c9.d20241201 depends on types-all, we can
      conclude that giga-connectome[dev]==0.1.dev198+gf6e41c9.d20241201 cannot be used.
      And because only giga-connectome[dev]==0.1.dev198+gf6e41c9.d20241201 is available
      and you require giga-connectome[dev], we can conclude that your requirements are
      unsatisfiable.
```

### Describe what you expected.

to be able to install the dev version, my local environment
