+++ .github/workflows/test.yml
@@ -158,7 +158,7 @@ jobs:
       if: ${{ contains(needs.check_skip_flags.outputs.head-commit-message, 'full_test') }} || ${{ github.event.pull_request.merged }}
       run: pytest --cov=giga_connectome --cov-report=xml --pyargs giga_connectome
 
-    - uses: codecov/codecov-action@v4
+    - uses: codecov/codecov-action@v5
       if: ${{ always() }}
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
