### 2024-12-02 bids-apps/giga_connectome#185 @jdkent
allow command line arguments to only use hyphens
### Your idea

allow --smoothing_fwhm and --participant_label to work with hypens as well (e.g.,`smoothing-fwhm` and `--participant-label`)
