### 2024-12-02 bids-apps/giga_connectome#186 @jdkent
memory multiplier of input
I'm curious what the peak memory consumption for an input file is. If a bold file is 1gb, will the peak memory being 2gb/5gb/10g? Are there large variables being held in memory longer than they need to be?
