### 2024-12-02 #7061 @jdkent
I have a few nitpicks as issues in the repository, but I do not see those as a reason to hold back the publication of this software in JOSS.
