### 2024-12-02 #7061 @mstimberg
Many thanks for your review, @jdkent. Just for reference, here the links to the issues you opened: bids-apps/giga_connectome#184, bids-apps/giga_connectome#185, bids-apps/giga_connectome#186

As you said, these do not seem to be major issues, but I'd suggest that the authors have a look at them and see whether they can be addressed easily.
