### 2024-12-03 8a38385
rephrase the intro
