+++ docs/source/index.rst
@@ -6,13 +6,15 @@
 Welcome to giga_connectome's documentation!
 ===========================================
 
-Functional connectivity is a common approach in analysing resting state fMRI data. Python tool Nilearn
-provides utilities to extract, denoise time-series on a parcellation and compute functional connectivity.
-Currently there's no standalone one stop solution to generate connectomes from fMRIPrep outputs.
-This BIDS-app combines Nilearn, TemplateFlow to denoise the data and generate timeseries and functional
-connectomes directly from fMRIPrep outputs.
-The workflow comes with several built in denoising strategies and several choices of atlases
-(MIST, Schaefer 7 networks, DiFuMo, Harvard-Oxford).
+Functional connectivity is a common approach in analysing resting state fMRI data.
+The Python tool Nilearn provides utilities to extract and denoise time-series on a parcellation.
+Nilearn also has methods to compute functional connectivity.
+While Nilearn provides useful methods to generate connectomes,
+there is no standalone one stop solution to generate connectomes from fMRIPrep outputs.
+Giga_connectome (a BIDS-app!) combines Nilearn and TemplateFlow to denoise the data, generate timeseries,
+and most critically giga_connectome generates functional connectomes directly from fMRIPrep outputs.
+The workflow comes with several built-in denoising strategies and
+there are several choices of atlases (MIST, Schaefer 7 networks, DiFuMo, Harvard-Oxford).
 Users can customise their own strategies and atlases using the configuration json files.
 
 .. toctree::
