### 2025-01-10 96ed715
[DOC] some edits of the docs (#187)
* rephrase the intro

* add instructions on adding ssh key
