### 2025-01-28 6e8f072
FIX mypy type on mask array for sanity check
