### 2025-01-28 bd96b43
FIX underscore -> hyphen in CLI
Closes #185
