### 2025-01-28 bids-apps/giga_connectome#185 @htwangtw
Thanks this is a great catch - I have changed all the underscores to hyphens in #190
However I am keeping `bids_dir`, `output_dir`, `analysis_level` - this is specified in BIDS app 1.0.
I know that there is a new standard coming out and believe it is the best to treat that as a separate issue.
