### 2025-01-28 5b81bd2
FIX use types-setuptools as types-all is deprecated
