### 2025-01-28 bids-apps/giga_connectome#184 @htwangtw
I have updated types-all to types-setuptools. I could now install the dev environment with no issues. Please let me know if this works for you now!
