### 2025-01-28 5f88615
Update README.md for all contributor bots
Closes #161
