+++ README.md
@@ -49,6 +49,17 @@ If not, [open up a new issue](https://github.com/bids-apps/giga_connectome/issue
 You can review open [issues]((https://github.com/bids-apps/giga_connectome/issues)) that we are looking for help with.
 If you submit a new pull request please be as detailed as possible in your comments.
 
+## Contributors
+
+<!-- ALL-CONTRIBUTORS-LIST:START - Do not remove or modify this section -->
+<!-- prettier-ignore-start -->
+<!-- markdownlint-disable -->
+
+<!-- markdownlint-restore -->
+<!-- prettier-ignore-end -->
+
+<!-- ALL-CONTRIBUTORS-LIST:END -->
+
 ## Acknowledgements
 
 If you use nilearn, please cite the corresponding paper: Abraham 2014,
