### 2025-01-28 739c89a
DOCS add instruction for all contributors bot
