### 2025-01-28 f154006
Add all contributor bots (#192)
* Update README.md for all contributor bots

Closes #161

* ENH add badge

* DOCS add instruction for all contributors bot

* DOCS update instructions

* ADD .all-contributorsrc
