### 2025-01-28 bids-apps/giga_connectome#161 @htwangtw
@all-contributors please add @clarkenj for user testing, example, bug reports. please add @Remi-Gau for infrastructure, maintenance. please add @pbellec for idea,financial.
please add @shnizzedy for bug reports. please add @jdkent for bug reports and documentation. please add @emullier for bug reports.
