### 2025-01-28 1261bea
allow hyphen or underscore for the args changed
