### 2025-01-28 fc18fce
FIX QD contribution (#197)
