### 2025-01-28 1ed16af
Fix/contributor (#198)
* FIX QD contribution

* ADD NC
