### 2025-01-28 9376867
docs: add Remi-Gau as a contributor for infra, and maintenance (#199)
* docs: update README.md

* docs: update .all-contributorsrc

---------

Co-authored-by: allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
