### 2025-01-28 8875c6d
docs: add shnizzedy as a contributor for bug (#202)
* docs: update README.md

* docs: update .all-contributorsrc

---------

Co-authored-by: allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
