### 2025-01-28 020135c
docs: add emullier as a contributor for bug (#203)
* docs: update README.md

* docs: update .all-contributorsrc

---------

Co-authored-by: allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
