### 2025-01-29 9a1dd33
FIX underscore -> hyphen in CLI (#190)
* FIX underscore -> hyphen in CLI

Closes #185

* FIX make sure test fits the same convention

* allow hyphen or underscore for the args changed
