### 2025-01-30 #7061 @mstimberg
:wave: Hi and a belated Happy New Year (or a not so belated Chinese New Year for all those who celebrate) 🎆 !

@emullier could you please give us an update:
> 👋 @emullier did you find time to have another look at the software? Please note that the authors have replied in the two issues you opened (https://github.com/bids-apps/giga_connectome/issues/171 and https://github.com/bids-apps/giga_connectome/issues/171).

@jdkent I think that @htwangtw addressed the issues that you opened, could you please have a look whether these solutions/answers work for you and update your checklist if necessary?

Thanks 🙏 !
