### 2025-01-30 a4f9f95
ADD atlas set up instructions
