### 2025-02-05 685b041
Added detailed instruction to run the container with test data
