+++ docs/source/usage.md
@@ -9,7 +9,89 @@
    :func: global_parser
 ```
 
-## Using customised configuration files for denoising strategy and atlas
+## General usage
+
+Here is an example of using the preset denoise strategies and atlases.
+
+### Download the example dataset
+
+We created a downsampled version of Open Neuro dataset
+[ds000017](https://openneuro.org/datasets/ds000017/versions/00001)
+hosted on [Zenodo page](https://zenodo.org/records/8091903).
+You can download this dataset to understand how to use different options of `giga_connectome`.
+
+You can download the data to any preferred location.
+For the purpose of this tutorial, we will save the downloaded dataset under `giga_connectome/test_data`
+in your home directory (`~/giga_connectome/test_data`).
+
+Here is the download instruction for Linux/Mac users:
+
+```bash
+mkdir -p ~/giga_connectome/test_data
+wget --retry-connrefused \
+    --waitretry=5 \
+    --read-timeout=20 \
+    --timeout=15 \
+    -t 0 \
+    -q \
+    -O ~/giga_connectome/test_data/ds000017.tar.gz \
+    "https://zenodo.org/record/8091903/files/ds000017-fmriprep22.0.1-downsampled-nosurface.tar.gz?download=1"
+tar -xzf ~/giga_connectome/test_data/ds000017.tar.gz -C ~/giga_connectome/test_data/
+rm ~/giga_connectome/test_data/ds000017.tar.gz
+```
+
+Alternatively, you can go to the [Zenodo page](https://zenodo.org/records/8091903) and download
+`ds000017-fmriprep22.0.1-downsampled-nosurface.tar.gz` and uncompress it.
+
+Now you will find the dataset at `~/giga_connectome/test_data/ds000017-fmriprep22.0.1-downsampled-nosurface`.
+Under this directory, you will find a full fMRIPrep output layout of two subjects.
+
+### Running `giga_connectome` with container
+
+Given that you have already [installed the container](./installation.md), you can run `giga_connectome` with
+[apptainer](https://apptainer.org/) as follow:
+
+```bash
+DATA=~/giga_connectome/test_data/ds000017-fmriprep22.0.1-downsampled-nosurface
+mkdir -p outputs/
+
+apptainer run \
+    --bind ${DATA}:/inputs \
+    --bind ./outputs:/outputs \
+    --bind ./outputs/atlases:/atlases \
+    giga_connectome \
+    /inputs \
+    /outputs \
+    participant \
+    -a /atlases \
+    --atlas Schaefer20187Networks \
+    --denoise-strategy simple \
+    --reindex-bids
+```
+
+For Docker:
+
+```bash
+DATA=${HOME}/giga_connectome/test_data/ds000017-fmriprep22.0.1-downsampled-nosurface
+mkdir -p outputs/
+
+docker run --rm \
+    -v ${DATA}:/test_data \
+    -v ./outputs:/outputs \
+    -v ./outputs/atlases:/atlases \
+    bids/giga_connectome \
+    /test_data \
+    /outputs \
+    participant \
+    -a /atlases \
+    --atlas Schaefer20187Networks \
+    --denoise-strategy simple \
+    --reindex-bids
+```
+
+Now you can navigate the outputs under `outputs`
+
+## Advanced: Using customised configuration files for denoising strategy and atlas
 
 Aside from the preset strategies and atlases, the users can supply their own for further customisation.
 We encourage the users to use the container version of the BIDS app, hence all documentation below will relect on the usage of the container.
@@ -118,7 +200,7 @@ Example:
 }
 ```
 
-See examples in [`giga_connectome/data`](https://github.com/bids-apps/giga_connectome/tree/main/giga_connectome/data).
[…]
