### 2025-02-05 bids-apps/giga_connectome#171 @htwangtw
@emullier Can you verify this issue is resolved with the current main branch, you pull the unstable version of the container to verify this
`docker pull bids/giga_connectome:unstable`
Further more we added more detailed documentations in #205
