### 2025-02-05 bids-apps/giga_connectome#166 @htwangtw
#205 added an real example using the test dataset in the usage page
