+++ .pre-commit-config.yaml
@@ -29,7 +29,7 @@ repos:
         args: ["--skip=*.svg"]
         exclude: 'paper/.*'
 -   repo: https://github.com/PyCQA/flake8
-    rev: 7.1.1
+    rev: 7.1.2
     hooks:
     -   id: flake8
 -   repo: https://github.com/hadolint/hadolint
