### 2025-04-14 #7061 @htwangtw
Hi @mstimberg do we have any update on either waiving COI or an alternative approach?
