### 2025-04-28 #7061 @mstimberg
Hi @htwangtw I've started looking at the software and was wondering which version you'd consider the one to be reviewed. The latest development version in the repository (corresponding to the `unstable` container I assume)? Also, I've seen that your PR bids-apps/giga_connectome#205 addresses earlier reviewer comments – is there a reason why you did not merge it yet?
