### 2025-04-29 bids-apps/giga_connectome#186 @htwangtw
Some recent PRs has addressed the issue in the upstream library: https://github.com/nilearn/nilearn/pull/5128
Close for now as this is a big issue that probably is better to revisit in the future
