### 2025-04-29 b7970d8
Improve documention for both contributors and user (#205)
* ADD atlas set up instructions

* Added detailed instruction to run the container with test data
