### 2025-04-29 #7061 @htwangtw
Thank you @mstimberg !!

> Hi [@htwangtw](https://github.com/htwangtw) I've started looking at the software and was wondering which version you'd consider the one to be reviewed. The latest development version in the repository (corresponding to the `unstable` container I assume)?

There was one release between the submission and finding reviewers. For simplicity, can we review the current version instead of the submitted version? All changes related to this review are in the `unstable` container indeed.

> Also, I've seen that your PR [bids-apps/giga_connectome#205](https://github.com/bids-apps/giga_connectome/pull/205) addresses earlier reviewer comments – is there a reason why you did not merge it yet?

I would like the original reviewer to have a look before merging, since they are now inactive, I am merging it now.
