### 2025-04-30 bids-apps/giga_connectome#211 @mstimberg
[BUG] Error for `--denoise-strategy icaaroma`
### What version of the bids app were you using?

Docker container `bids/giga_connectome:unstable`
Version: 0.5.1.dev57+ga5f209d

### Describe your problem in detail.

When I run the example from https://giga-connectome.readthedocs.io/en/latest/usage.html#running-giga-connectome-with-container but replace `--denoise-strategy simple` by `--denoise-strategy icaaroma`, I get the following error:
```pytb
Traceback (most recent call last):
  File "/usr/local/bin/giga_connectome", line 8, in <module>
    sys.exit(main())
  File "/usr/local/lib/python3.9/site-packages/giga_connectome/run.py", line 165, in main
    workflow(args)
  File "/usr/local/lib/python3.9/site-packages/giga_connectome/workflow.py", line 82, in workflow
    subject_mask_nii, subject_seg_niis = generate_gm_mask_atlas(
  File "/usr/local/lib/python3.9/site-packages/giga_connectome/mask.py", line 38, in generate_gm_mask_atlas
    subject, _, _ = utils.parse_bids_name(masks[0].path)
IndexError: list index out of range
```

[Opened as part of JOSS review:  openjournals/joss-reviews#7061]

### What command did you run?

```
$ docker run --rm \                                                                                                                                                  18:51
      -v {$DATA}:/test_data \
      -v ./outputs:/outputs \
      -v ./outputs/atlases:/atlases \
      bids/giga_connectome:unstable \
      /test_data \
      /outputs_ica_aroma \
      participant \
      --atlases-dir /atlases \
      --atlas Schaefer20187Networks \
      --denoise-strategy icaaroma \
      --reindex-bids
```

### Describe what you expected.

_No response_
