### 2025-04-30 bids-apps/giga_connectome#212 @mstimberg
[doc] suggestions for doc improvements
With respect to https://giga-connectome.readthedocs.io/en/latest/ (commit b7970d8a), I have a few suggestions for the "Usage Notes" page:
- similar to the changes in #166, I think it would be good to have the "advanced" examples to also use actual values instead of the hypothetical ones.
- In the documentation for the command line arguments, line breaks seem to be not handled correctly, leading to some missing spaces and at least one broken link
- "See examples in `giga_connectome/data/atlas/`" has a broken link
- The example should use `Schaefer2018` instead of `Schaefer20187Networks` to avoid the deprecation warning.
- The last example uses `-s /data/atlases` instead of **`-a`**

Regarding the README file:
- It would be helpful if the "How to contribute" section linked to the Contribution section of the documentation (which is nicely detailed)
- In the Acknowledgements, "If you use nilearn, please cite the corresponding paper:" reads a bit odd in the context of this README – do the authors suggest users of Giga-Connectome should cite `nilearn` (given that it uses `nilearn` under the hood)?

Not really a documentation issue, but I noticed that the LICENSE file states `Copyright (c) 2022 SIMEXP` which should probably be updated?

[Opened as part of JOSS review: openjournals/joss-reviews#7061]
