### 2025-04-30 bids-apps/giga_connectome#213 @mstimberg
[BUG] Typo in command line help
### What version of the bids app were you using?

Docker container `bids/giga_connectome:unstable`
Version: 0.5.1.dev57+ga5f209d

### Describe your problem in detail.

The command line help states:
```
  -w WORK_DIR, --work-dir WORK_DIR
                        This argument is deprecated. Please use --atlas-dir instead.
```
but the correct argument is called `--atlases-dir` not `--atlas-dir`.

[Opened as part of JOSS review:  openjournals/joss-reviews#7061]

### What command did you run?

_No response_

### Describe what you expected.

_No response_
