### 2025-05-02 d80dcf2
DOCS/FIX typo in docstrings
