### 2025-05-02 7940b2f
DOCS Update nilearn citation details; link to contribution docs
