+++ giga_connectome/run.py
@@ -82,7 +82,7 @@ def global_parser() -> argparse.ArgumentParser:
         "--atlases-dir",
         action="store",
         type=Path,
-        default=Path("atlases").absolute(),
+        default=Path("atlases"),
         help="Path where subject specific segmentations are stored.",
     )
     parser.add_argument(
