+++ docs/source/usage.md
@@ -122,19 +122,12 @@ In a `json` file, define the customised strategy in the following format:
 
 See examples in [`giga_connectome/data/denoise_strategy`](https://github.com/bids-apps/giga_connectome/tree/main/giga_connectome/data/denoise_strategy).
 
-2. Mount the path to the configuration file to the container and pass the **mounted path** to `--denoise-strategy`.
-
-An example using Apptainer (formerly known as Singularity):
+In the following section, we will use this customised denoising strategy (5 aCompCor components, 6 motion parameters, global signal).
+Run the following section in bash to create the denoising configuration file.
 
 ```bash
-# create denoising strategy
-mkdir custome_denoise
-DATA=${HOME}/giga_connectome/test_data/ds000017-fmriprep22.0.1-downsampled-nosurface
-OUTPUT_DIR=${HOME}/custome_denoise/connectom_output
-ATLASES_DIR=${HOME}/custome_denoise/atlas_output
-DENOISE_CONFIG=${HOME}/custome_denoise/denoise_config.json
-
-GIGA_CONNECTOME=${HOME}/giga-connectome.simg
+mkdir ${HOME}/customised_denoise
+DENOISE_CONFIG=${HOME}/customised_denoise/denoise_config.json
 
 # create denoise file
 cat << EOF > ${DENOISE_CONFIG}
@@ -151,6 +144,21 @@ cat << EOF > ${DENOISE_CONFIG}
     }
 }
 EOF
+```
+
+2. Mount the path to the configuration file to the container and pass the **mounted path** to `--denoise-strategy`.
+
+An example using Apptainer, with data downloaded as described in the [previous section](#download-the-example-dataset) :
+
+```bash
+# create denoising strategy
+mkdir ${HOME}/customised_denoise/outputs
+DATA=${HOME}/giga_connectome/test_data/ds000017-fmriprep22.0.1-downsampled-nosurface
+OUTPUT_DIR=${HOME}/customised_denoise/outputs
+ATLASES_DIR=${HOME}/customised_denoise/outputs/atlas
+DENOISE_CONFIG=${HOME}/customised_denoise/denoise_config.json
+
+GIGA_CONNECTOME=${HOME}/giga-connectome.simg  # assuming the container is created in $HOME
 
 apptainer run \
     --bind ${DATA}:/data/input \
@@ -161,6 +169,7 @@ apptainer run \
     -a /data/atlases \
     --atlas Schaefer2018 \
     --denoise-strategy /data/denoise_config.json \
+    --reindex-bids \
     /data/input \
     /data/output \
     participant
@@ -170,29 +179,11 @@ For Docker:
 
 ```bash
 # create denoising strategy
-mkdir custome_denoise
+mkdir ${HOME}/customised_denoise/outputs
 DATA=${HOME}/giga_connectome/test_data/ds000017-fmriprep22.0.1-downsampled-nosurface
-OUTPUT_DIR=${HOME}/custome_denoise/connectom_output
-ATLASES_DIR=${HOME}/custome_denoise/atlas_output
-DENOISE_CONFIG=${HOME}/custome_denoise/denoise_config.json
-
-GIGA_CONNECTOME=${HOME}/giga-connectome.simg
-
-# create denoise file
-cat << EOF > ${DENOISE_CONFIG}
-{
-    "name": "custom_compcor",
-    "function": "load_confounds",
-    "parameters": {
-        "strategy": ["high_pass", "motion", "compcor"],
-        "motion": "basic",
-        "n_compcor": 5,
-        "compcor": "anat_combined",
-        "global_signal": "basic",
[…]
