### 2025-05-05 bids-apps/giga_connectome#212 @htwangtw
@mstimberg Thank you for these are great points to improve on!
If you feel all points are addressed in PR#215, then we can merge it
