### 2025-05-05 bids-apps/giga_connectome#211 @htwangtw
I can reproduce the issue and found the issue in `giga_connectome/utils.py:get_bids_images`

The BIDS filter does not account for the case for ICA-AROMA.

I will fix this part of the code and add unit test on this issue
