### 2025-05-06 0f136f3
FIX create BIDS filter for ICA-AROMA
