### 2025-05-06 813e868
FIX ICA AROMA
