+++ giga_connectome/tests/test_denoise.py
@@ -31,3 +31,17 @@ def test_denoise_nifti_voxel():
     testing.assert_almost_equal(
         meta_data["MeanFramewiseDisplacement"], 0.107, decimal=3
     )
+
+    img_file = resource_filename(
+        "giga_connectome",
+        "data/test_data/ds000017-fmriprep22.0.1-downsampled-nosurface/sub-1/ses-timepoint1/func/sub-1_ses-timepoint1_task-probabilisticclassification_run-01_space-MNI152NLin6Asym_desc-smoothAROMAnonaggr_bold.nii.gz",
+    )
+    strategy = get_denoise_strategy("icaaroma")
+    meta_data = denoise_meta_data(
+        strategy=strategy,
+        img=img_file,
+    )
+    assert len(meta_data["ConfoundRegressors"]) == 15
+    assert len(meta_data["ICAAROMANoiseComponents"]) == 9
+    assert meta_data["NumberOfVolumesDiscardedByMotionScrubbing"] == 0
+    assert meta_data["NumberOfVolumesDiscardedByNonsteadyStatesDetector"] == 2
