### 2025-05-06 ebf23ff
FIX calculation of non steady state
