### 2025-05-06 236ab60
FIX incorrect number of confound count in test; the original number merges ICA aroma components in there
