### 2025-05-06 a6e64f3
JOSS - Improvement on advance usages examples (#215)
* DOCS Update nilearn citation details; link to contribution docs

* FIX borken links and outdated arguments in the usage notes

* DOCS make the line break reads more comfortably

* DOCS try see if this works.

* DOCS clarify it's in working dir

* it works

* DOCS advanced usage with working exaple

* MAINT update changelog

* MAINT update license to BIDS Apps
