+++ giga_connectome/run.py
@@ -82,7 +82,7 @@ def global_parser() -> argparse.ArgumentParser:
         "--atlases-dir",
         action="store",
         type=Path,
-        default=Path("atlases").absolute(),
+        default=Path("./atlases"),
         help="Path where subject specific segmentations are stored.",
     )
     parser.add_argument(
@@ -125,11 +125,11 @@ def global_parser() -> argparse.ArgumentParser:
     parser.add_argument(
         "--bids-filter-file",
         type=Path,
-        help="A JSON file describing custom BIDS input filters using PyBIDS."
+        help="A JSON file describing custom BIDS input filters using PyBIDS. "
         "We use the same format as described in fMRIPrep documentation: "
         "https://fmriprep.org/en/latest/faq.html#"
-        "how-do-i-select-only-certain-files-to-be-input-to-fmriprep"
-        "However, the query filed should always be 'bold'",
+        "how-do-i-select-only-certain-files-to-be-input-to-fmriprep "
+        "\nHowever, the query filed should always be 'bold'",
     )
     parser.add_argument(
         "--calculate-intranetwork-average-correlation",
