+++ docs/source/usage.md
@@ -64,7 +64,7 @@ apptainer run \
     /outputs \
     participant \
     -a /atlases \
-    --atlas Schaefer20187Networks \
+    --atlas Schaefer2018 \
     --denoise-strategy simple \
     --reindex-bids
 ```
@@ -84,7 +84,7 @@ docker run --rm \
     /outputs \
     participant \
     -a /atlases \
-    --atlas Schaefer20187Networks \
+    --atlas Schaefer2018 \
     --denoise-strategy simple \
     --reindex-bids
 ```
@@ -122,34 +122,91 @@ In a `json` file, define the customised strategy in the following format:
 
 See examples in [`giga_connectome/data/denoise_strategy`](https://github.com/bids-apps/giga_connectome/tree/main/giga_connectome/data/denoise_strategy).
 
-2. Mount the path to the configuration file to the container and pass the **mounted path** to `--denoise-strategy`.
-
-An example using Apptainer (formerly known as Singularity):
+In the following section, we will use this customised denoising strategy (5 aCompCor components, 6 motion parameters, global signal).
+Run the following section in bash to create the denoising configuration file.
 
 ```bash
-FMRIPREP_DIR=/path/to/fmriprep_output
-OUTPUT_DIR=/path/to/connectom_output
-ATLASES_DIR=/path/to/atlases
-DENOISE_CONFIG=/path/to/denoise_config.json
+mkdir ${HOME}/customised_denoise
+DENOISE_CONFIG=${HOME}/customised_denoise/denoise_config.json
 
-GIGA_CONNECTOME=/path/to/giga-connectome.simg
+# create denoise file
+cat << EOF > ${DENOISE_CONFIG}
+{
+    "name": "custom_compcor",
+    "function": "load_confounds",
+    "parameters": {
+        "strategy": ["high_pass", "motion", "compcor"],
+        "motion": "basic",
+        "n_compcor": 5,
+        "compcor": "anat_combined",
+        "global_signal": "basic",
+        "demean": true
+    }
+}
+EOF
+```
+
+2. Mount the path to the configuration file to the container and pass the **mounted path** to `--denoise-strategy`.
+
+An example using Apptainer, with data downloaded as described in the [previous section](#download-the-example-dataset) :
+
+```bash
+# create denoising strategy
+mkdir ${HOME}/customised_denoise/outputs
+DATA=${HOME}/giga_connectome/test_data/ds000017-fmriprep22.0.1-downsampled-nosurface
+OUTPUT_DIR=${HOME}/customised_denoise/outputs
+ATLASES_DIR=${HOME}/customised_denoise/outputs/atlas
+DENOISE_CONFIG=${HOME}/customised_denoise/denoise_config.json
+
+GIGA_CONNECTOME=${HOME}/giga-connectome.simg  # assuming the container is created in $HOME
 
 apptainer run \
-    --bind ${FMRIPREP_DIR}:/data/input \
+    --bind ${DATA}:/data/input \
     --bind ${OUTPUT_DIR}:/data/output \
     --bind ${ATLASES_DIR}:/data/atlases \
     --bind ${DENOISE_CONFIG}:/data/denoise_config.json \
     ${GIGA_CONNECTOME} \
     -a /data/atlases \
+    --atlas Schaefer2018 \
     --denoise-strategy /data/denoise_config.json \
+    --reindex-bids \
     /data/input \
     /data/output \
     participant
 ```
 
+For Docker:
+
+```bash
+# create denoising strategy
[…]
