### 2025-05-06 e0821ae
FIX ICA-AROMA implementation (#216)
* FIX create BIDS filter for ICA-AROMA

* FIX ICA AROMA

* FIX calculation of non steady state

* FIX incorrect number of confound count in test; the original number merges ICA aroma components in there
