+++ giga_connectome/utils.py
@@ -20,8 +20,52 @@ from rich.progress import (
 
 from giga_connectome import __version__
 from giga_connectome.logger import gc_logger
+from giga_connectome.denoise import is_ica_aroma, STRATEGY_TYPE
+
 
 gc_log = gc_logger()
+
+
+def prepare_bidsfilter_and_template(
+    strategy: STRATEGY_TYPE,
+    user_bids_filter: None | dict[str, dict[str, str]],
+) -> tuple[str, dict[str, dict[str, str]]]:
+    """
+    Prepare the template and BIDS filters for ICA AROMA.
+    This solution only applies to fMRIPrep version < 23.1.0.
+    For later versions, we will wait for updates in the upstream library
+    nilearn to support the new layoput for ICA AROMA.
+
+    Parameters
+    ----------
+    strategy : str
+        The denoising strategy.
+    bids_filter_file : Path
+        The path to the BIDS filter file.
+
+    Returns
+    -------
+    tuple[str, None | dict[str, dict[str, str]]]
+        The template and BIDS filters.
+    """
+    user_bids_filter = check_filter(user_bids_filter)
+    if is_ica_aroma(strategy):
+        template = "MNI152NLin6Asym"
+        bids_filters = {  # this only applies to fMRIPrep version < 23.1.0
+            "bold": {
+                "desc": "smoothAROMAnonaggr",
+                "space": "MNI152NLin6Asym",
+            },
+            "mask": {  # no brain mask for MNI152NLin6Asym, use the closest
+                "space": "MNI152NLin2009cAsym",
+            },
+        }
+        if user_bids_filter:
+            for suffix, entities in user_bids_filter.items():
+                bids_filters[suffix].update(entities)
+        return template, bids_filters
+    else:
+        return "MNI152NLin2009cAsym", user_bids_filter
 
 
 def get_bids_images(
@@ -35,6 +79,7 @@ def get_bids_images(
     Apply BIDS filter to the base filter we are using.
     Modified from fmripprep
     """
+
     bids_filters = check_filter(bids_filters)
 
     layout = BIDSLayout(
@@ -50,7 +95,6 @@ def get_bids_images(
         "return_type": "object",
         "subject": subjects,
         "session": Query.OPTIONAL,
-        "space": template,
         "task": Query.ANY,
         "run": Query.OPTIONAL,
         "extension": ".nii.gz",
@@ -58,11 +102,13 @@ def get_bids_images(
     queries = {
         "bold": {
             "desc": "preproc",
+            "space": template,
             "suffix": "bold",
             "datatype": "func",
         },
         "mask": {
             "suffix": "mask",
+            "space": "MNI152NLin2009cAsym",
             "datatype": "func",
         },
     }
@@ -114,6 +160,18 @@ def _filter_pybids_none_any(
 
 
 def parse_bids_filter(value: Path) -> None | dict[str, dict[str, str]]:
+    """Parse a BIDS filter json file.
+    Parameters
+    ----------
+    value : Path
+        Path to the BIDS json file.
+
+    Returns
+    -------
+    dict
+        Dictionary of BIDS filters.
+    """
+
     from json import JSONDecodeError, loads
 
     if not value:

[…]
