### 2025-05-06 96820d8
docs: add mstimberg as a contributor for userTesting, doc, and bug (#218)
* docs: update README.md

* docs: update .all-contributorsrc

---------

Co-authored-by: allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
