### 2025-05-06 be177c0
missed changelog item
