### 2025-05-07 bids-apps/giga_connectome#220 @mstimberg
[doc] a few more minor suggestions
Thanks for all the changes in #215, I can confirm that the updated "advanced instructions" work well.

I have a few more minor nitpicking points regarding https://giga-connectome.readthedocs.io/en/latest/:
- https://giga-connectome.readthedocs.io/en/latest/usage.html#atlas : Warning box uses `nilear` instead of `nilearn`
- The following paragraph has a spurious `mkdir`:
  > In the following section, we will generate an atlas on the test data using nilearn.regions.Parcellations with ward clustering method. Run the following python code to generate the atlas and save to mkdir ${HOME}/customised_atlas/templateflow in templateflow convention.
- The very last docker example on the page starts with `# create denoising strategy` but that's not done here, the bash script before to generate `ward_config.json` also has a `# create denoise file` comment
- in the contribution section, the "demo" at https://giga-connectome.readthedocs.io/en/latest/contributing.html#running-the-demo does not work since it uses an unsupported (removed?) `--standardize` flag. Also, it gives a confusing error where it tries to interpret the output dir as the analysis_level:
  > giga_connectome: error: argument analysis_level: invalid choice: 'giga_connectome/data/test_data/output' (choose from 'participant')

[Opened as part of JOSS review: openjournals/joss-reviews#7061
