### 2025-05-07 bids-apps/giga_connectome#221 @mstimberg
[docs/paper] clarify BEP 017 status
From the paper and the docs at https://giga-connectome.readthedocs.io/en/latest/outputs.html it wasn't quite clear to me that BEP 017 is still a draft proposal that is "in-progress". Maybe that could be clarified.
Also, from a cursory look at the spec, I got the impression that each `_relmat.tsv` would have an accompanying `_relmat.json` sidecar file, but GigaConnectome creates a single `_relmat.json` file at the output root – would you consider this being in line with the spec? Apologies if that question results from my very superficial understanding of BIDS…

[Opened as part of JOSS review: openjournals/joss-reviews#7061
