### 2025-05-08 bids-apps/giga_connectome#221 @htwangtw
Not at all!

- It's encouraged to adopt BIDS extension proposals, however few in the community are doing this. I think it's reasonable to have some extra lines of descriptions to let user know the status and plans to keeping up with the updates

- As for `_relmat.json` since there are one configuration for each atlas, this would be considered a common element across datasets, hence the `_relmat.json` is at root level. Similar usecase include how entity `task-*.json` for fMRI data is not in the meta data with the `.nii.gz` file but at dataset level.

Let me know if you have more questions!
