### 2025-05-08 cdd09fe
Minor edits on docs (#222)
* small errors

* clarification of BEP017

* minor formatting edits for the paper
