### 2025-05-09 bids-apps/giga_connectome#221 @mstimberg
Ok, thanks, that works for me. Note that there is a small mistake in the first sentence of the "Outputs" page: it ends with "and ."
