### 2025-05-12 #7061 @mstimberg
Hi @htwangtw. I made another small PR (bids-apps/giga_connectome#223) to fix a few capitalization issues in the paper/bibliography.
In addition, could you please make the affiliations a bit more uniform? In particular, decide whether you prefer "University of Montreal" or "Université de Montréal". Also, please try to use a "Department, University, City, Country" format everywhere. But also note that while we prefer this style, if an author's institution imposes a specific style that conflicts with this, please note this here in a comment and we'll happily accept a slightly diverging format.
