### 2025-05-12 08b8192
Editing fixes for JOSS paper (#223)
* JOSS - Minor editing changes to paper

* JOSS - capitalization fixes in bibliography
