+++ paper/paper.md
@@ -39,7 +39,7 @@ bibliography: paper.bib
 
 # Summary
 
-Researchers perform two steps before Functional magnetic resonance imaging (fMRI) data analysis:
+Researchers perform two steps before functional magnetic resonance imaging (fMRI) data analysis:
 standardised preprocessing and customised denoising.
 `fMRIPrep` [@fmriprep\; RRID:SCR_016216],
 a popular software in the neuroimaging community, is a common choice for preprocessing.
@@ -48,14 +48,14 @@ The present software, `giga-connectome`,
 is a Brain Imaging Data Structure [BIDS\; @bids\; RRID:SCR_016124]
 compliant container image that aims to perform these steps as well as extract time series signals and generate connectomes for machine learning applications.
 All these steps are implemented with functions from `nilearn` [@Nilearn\; RRID:SCR_001362],
-a python library for machine learning in neuroimaging.
+a Python library for machine learning in neuroimaging.
 
 The tool performs smoothing, denoising, and standardisation on voxel level data.
 Smoothing is implemented with a 5mm full width at half maximum kernel and the user can change the kernel size based on the voxel size of their fMRI data.
 For the denoising step, we built the workflow closely aligned with the design choice of `fmriprep`
 and worked with the `fmriprep` developers while implementing a key Application Programming Interface (API),
 `load_confounds`, implemented in the software library `nilearn`.
-The tool provides some preset strategies based on @wang_continuous_2024 and the current long-term support release `fMRIPrep`.
+The tool provides some preset strategies based on @wang_continuous_2024 and the current long-term support release of `fMRIPrep`.
 Users can implement their own strategy using configuration files to directly interact with the `load_confounds` API.
 The details of the process can be found in the [user documentation](https://giga-connectome.readthedocs.io/en/latest/workflow.html).
 Finally the data is standardised as z-scores.
@@ -73,10 +73,10 @@ We aim to include more default atlases when they are included in the `templatefl
 The time series extraction is implemented with `nilearn` objects `NiftiLabelsMasker` and `NiftiMapsMasker`.
 The generated time series are used to construct connectomes calculated as Pearson's correlation with `nilearn` object `ConnectivityMeasure`.
 
-Finally the saved time series and connectomes follow the format of
+Finally the saved time series and connectomes follow the format of the
 [BIDS-connectome specification](https://bids.neuroimaging.io/bep017).
 Users can follow the specification to interpret and retrieve the relevant results.
-The coverage of the atlas is also included as a html visual report, provided by `nilearn.masker` for users to examine the quality of the atlas coverage.
+The coverage of the atlas is also included as an HTML visual report, provided by `nilearn.masker` for users to examine the quality of the atlas coverage.
 More information about the usage, workflow, and outputs can be found on the
 [official documentation](https://giga-connectome.readthedocs.io/en/latest/).
 
@@ -85,9 +85,9 @@ More information about the usage, workflow, and outputs can be found on the
 
 `giga-connectome` is created for large scale deployment on multiple `fMRIPrep` preprocessed neuroimaging datasets.
 We aimed to create a tool that is lightweight in terms of code base complexity, software dependencies, and command line interface (CLI).
-The current software follows the BIDS-apps API [@bidsapp] and is the first of its kind that creates outputs in the BIDS-connectome.
+The current software follows the BIDS-apps API [@bidsapp] and is the first of its kind that creates outputs following the BIDS-connectome specification.
 Both users and developers would benefit from the detailed definition of BIDS-apps API and BIDS-connectome for shared usage documentation and development guidelines.
[…]
