+++ paper/paper.bib
@@ -21,7 +21,7 @@
 }
 
 @article{bidsapp,
-	title = {{BIDS} apps: {Improving} ease of use, accessibility, and reproducibility of neuroimaging data analysis methods},
+	title = {{BIDS} apps: Improving ease of use, accessibility, and reproducibility of neuroimaging data analysis methods},
 	volume = {13},
 	issn = {1553-7358},
 	doi = {10.1371/journal.pcbi.1005209},
@@ -136,7 +136,7 @@
 }
 
 @article{schaefer_local-global_2018,
-	title = {Local-{Global} {Parcellation} of the {Human} {Cerebral} {Cortex} from {Intrinsic} {Functional} {Connectivity} {MRI}},
+	title = {Local-Global Parcellation of the Human Cerebral Cortex from Intrinsic Functional Connectivity MRI},
 	volume = {28},
 	issn = {1460-2199},
 	doi = {10.1093/cercor/bhx179},
@@ -152,7 +152,7 @@
 }
 
 @article{urchs_mist_2019,
-	title = {{MIST}: {A} multi-resolution parcellation of functional brain networks},
+	title = {{MIST}: A multi-resolution parcellation of functional brain networks},
 	volume = {1},
 	issn = {2515-5059},
 	shorttitle = {{MIST}},
@@ -190,7 +190,7 @@
 
 
 @article{HALFpipe,
-	title = {{ENIGMA} {HALFpipe}: {Interactive}, reproducible, and efficient analysis for resting-state and task-based {fMRI} data},
+	title = {{ENIGMA} {HALFpipe}: Interactive, reproducible, and efficient analysis for resting-state and task-based {fMRI} data},
 	volume = {43},
 	issn = {1097-0193},
 	doi = {10.1002/hbm.25829},
@@ -206,7 +206,7 @@
 }
 
 @misc{cpac,
-	title = {Moving {Beyond} {Processing} and {Analysis}-{Related} {Variation} in {Neuroscience}},
+	title = {Moving Beyond Processing and Analysis-Related Variation in Neuroscience},
 	doi = {10.1101/2021.12.01.470790},
 	language = {en},
 	urldate = {2024-03-28},
@@ -231,7 +231,7 @@
 }
 
 @misc{xcp-d,
-	title = {{XCP}-{D}: {A} {Robust} {Pipeline} for the post-processing of {fMRI} data},
+	title = {{XCP}-{D}: A Robust Pipeline for the post-processing of {fMRI} data},
 	shorttitle = {{XCP}-{D}},
 	doi = {10.1101/2023.11.20.567926},
 	language = {en},
@@ -243,7 +243,7 @@
 }
 
 @article{ABIDE,
-	title = {The {Autism} {Brain} {Imaging} {Data} {Exchange}: {Towards} {Large}-{Scale} {Evaluation} of the {Intrinsic} {Brain} {Architecture} in {Autism}},
+	title = {The Autism Brain Imaging Data Exchange: Towards Large-Scale Evaluation of the Intrinsic Brain Architecture in Autism},
 	volume = {19},
 	issn = {1359-4184},
 	doi = {10.1038/mp.2013.78},
@@ -258,7 +258,7 @@
 }
 
 @article{ukbiobank,
-	title = {{UK} {Biobank}: {An} {Open} {Access} {Resource} for {Identifying} the {Causes} of a {Wide} {Range} of {Complex} {Diseases} of {Middle} and {Old} {Age}},
+	title = {{UK} {Biobank}: An Open Access Resource for Identifying the Causes of a Wide Range of Complex Diseases of Middle and Old Age},
 	volume = {12},
 	issn = {1549-1676},
 	doi = {10.1371/journal.pmed.1001779},
@@ -272,7 +272,7 @@
 }
 
[…]
