### 2025-05-12 a065677
[JOSS] update affiliations (#224)
* [JOSS] update affiliations

* yaml formating
