### 2025-05-12 #7061 @mstimberg
Thanks for the update @htwangtw. Regarding the affiliations, I see one remaining issue: the affiliation "3
MIND team, INRIA, CEA, Université Paris-Saclay, Paris, France" does not appear to be linked to any author.
