### 2025-05-14 2742000
detail the scope of implemntation
