### 2025-05-14 57416ff
DOCS add fmriprep version details (#227)
* DOCS add fmriprep version details

* detail the scope of implemntation
