+++ CITATION.cff
@@ -9,12 +9,12 @@ message: "If you use this software, please cite it as below."
 
 repository-code: "https://github.com/bids-apps/giga_connectome"
 
+license: MIT
+
 contact:
   - email: htwangtw@gmail.com
     family-names: Wang
     given-names: Hao-Ting
-
-license: MIT
 
 authors:
   - family-names: Wang
@@ -27,4 +27,4 @@ authors:
   - family-names: Clarke
     given-names: Natasha
   - family-names: Bellec
-    given-names: Pierre
+    given-names: Lune
