### 2025-05-14 74b023d
REL 0.6.0 (#226)
* updating version numbers

* REL 0.6.0

* clarify license

* correct names
