### 2025-05-21 #7061 @htwangtw
> Thank you [@htwangtw](https://github.com/htwangtw). Could you please update the Zenodo archive so that it has the same title as the paper? Note that you can edit this without creating a new record/DOI.

@mstimberg It should now be resolved!
