### 2025-06-09 cc952d3
Fix typo in paper and update Lune's name
